Moved WPF ViewModel tests to their own folder.
This is in preparation for adding other types of tests.
